Разделит общий редьюсер на отдельные, добавит селекторы
