+++ src/app/router/index.tsx
@@ -2,24 +2,22 @@ import { useEffect } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import { PrivateRoute } from '@/app/router/private-route';
+import { checkLogin, getAuthorizationStatus } from '@/entities/user';
 import { Error404Page } from '@/pages/error';
 import { FavoritesPage } from '@/pages/favorites';
 import { LoginPage } from '@/pages/login';
 import { MainPage } from '@/pages/main';
 import { OfferPage } from '@/pages/offer';
 import { AppRoutes } from '@/shared/model/app-routes';
+import { AuthorizationStatus } from '@/shared/model/enums';
 import { Spinner } from '@/shared/ui/spinner';
 
-import { AuthorizationStatus } from '../../shared/model/enums';
-import { checkLogin } from '../store/model/async-thunks';
 import { useAppDispatch, useAppSelector } from '../store/model/hooks';
 
 export const AppRouter = () => {
   const dispatch = useAppDispatch();
 
-  const authorizationStatus = useAppSelector(
-    (state) => state.authorizationStatus,
-  );
+  const authorizationStatus = useAppSelector(getAuthorizationStatus);
 
   useEffect(() => {
     dispatch(checkLogin());
