+++ src/app/router/private-route/index.tsx
@@ -2,12 +2,10 @@ import { PropsWithChildren } from 'react';
 import { Navigate } from 'react-router-dom';
 
 import { useAppSelector } from '@/app/store/model/hooks';
-import { AuthorizationStatus } from '@/shared/model/enums';
+import { getIsAuthenticated } from '@/entities/user';
 
 export const PrivateRoute = ({ children }: PropsWithChildren) => {
-  const isAuthorized = useAppSelector(
-    (state) => state.authorizationStatus === AuthorizationStatus.AUTH,
-  );
+  const isAuthorized = useAppSelector(getIsAuthenticated);
 
   return isAuthorized ? children : <Navigate to={'/login'} />;
 };
