--- src/app/store/model/async-thunks/index.ts
@@ -1,10 +1,0 @@
-export { addOfferReview } from './add-offer-review';
-export { checkLogin } from './check-login';
-export { fetchFavoriteOffers } from './fetch-favorite-offers';
-export { fetchOffer } from './fetch-offer';
-export { fetchOfferReviews } from './fetch-offer-reviews';
-export { fetchOffers } from './fetch-offers';
-export { fetchOffersNearby } from './fetch-offers-nearby';
-export { login } from './login';
-export { logout } from './logout';
-export { setIsOfferFavorite } from './set-favorite-offer';
