--- src/app/store/model/async-thunks/logout.ts
@@ -1,18 +1,0 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-
-import { removeToken } from '@/app/api/lib/token';
-import { ApiRoutes } from '@/app/api/model/api-routes';
-import { AuthorizationStatus } from '@/shared/model/enums';
-
-import { setAuthorizationStatus, setUser } from '../actions';
-import { DispatchStateExtra } from '../types';
-
-export const logout = createAsyncThunk<void, undefined, DispatchStateExtra>(
-  'user/logout',
-  async (_arg, { dispatch, extra: api }) => {
-    await api.delete(ApiRoutes.LOGOUT);
-    dispatch(setAuthorizationStatus(AuthorizationStatus.NO_AUTH));
-    dispatch(setUser(undefined));
-    removeToken();
-  },
-);
