+++ src/app/store/model/types.ts
@@ -1,38 +1,13 @@
 import { AxiosInstance } from 'axios';
 
-import { components } from '@/../types/schema';
-import { City } from '@/entities/city';
-import { OfferMaximum, OfferPreview } from '@/entities/offer-card';
-import { CommentGet } from '@/entities/review';
-import { AuthInfo } from '@/entities/user';
-import { AuthorizationStatus, FetchStatus } from '@/shared/model/enums';
-
 import { store } from '..';
 
-export type State = {
-  city: City;
+export type State = ReturnType<typeof store.getState>;
 
-  user?: AuthInfo;
-  authorizationStatus: AuthorizationStatus;
-
-  offers?: OfferPreview[];
-  offersFetchStatus: FetchStatus;
-
-  offer?: OfferMaximum;
-  offerFetchStatus: FetchStatus;
-
-  reviews?: CommentGet[];
-  reviewsFetchStatus: FetchStatus;
-
-  activeOfferId?: OfferPreview['id'];
-};
+export type AppDispatch = typeof store.dispatch;
 
 export type DispatchStateExtra = {
   dispatch: AppDispatch;
   getState: () => State;
   extra: AxiosInstance;
 };
-
-export type AppDispatch = typeof store.dispatch;
-
-export type AuthResponse = Required<components['schemas']['AuthInfoWithToken']>;
