+++ src/entities/city/index.ts
@@ -1,3 +1,6 @@
 export { cities } from './model/consts';
 export { CityNames } from './model/types';
 export type { CityName, City } from './model/types';
+
+export * from './model/reducer';
+export * from './model/selectors';
