+++ src/entities/offer-card/index.ts
@@ -4,3 +4,7 @@ export type {
   OfferPreview as OfferPreview,
   OfferMaximum,
 } from './model/types';
+
+export * from './model/selectors';
+export * from './model/reducer';
+export * from './api/api-actions';
