+++ src/entities/review/index.ts
@@ -1,2 +1,6 @@
 export { Review } from './ui';
 export type { CommentGet, CommentPost } from './model/types';
+
+export * from './model/reducer';
+export * from './model/selectors';
+export * from './api/api-actions';
