+++ src/entities/user/index.ts
@@ -1,1 +1,5 @@
 export type { AuthInfo, User } from './model/types';
+
+export * from './model/selectors';
+export * from './model/reducer';
+export * from './api/api-actions';
