+++ src/pages/favorites/index.tsx
@@ -1,8 +1,12 @@
 import { useEffect } from 'react';
 
-import { clearOffers } from '@/app/store/model/actions';
-import { fetchFavoriteOffers } from '@/app/store/model/async-thunks';
 import { useAppDispatch, useAppSelector } from '@/app/store/model/hooks';
+import {
+  clearOffers,
+  fetchFavoriteOffers,
+  getOffers,
+  getOffersFetchStatus,
+} from '@/entities/offer-card';
 import { FavoriteOffersList } from '@/features/favorite-offers-list';
 import { Footer } from '@/features/footer';
 import { Header } from '@/features/header';
@@ -12,8 +16,8 @@ import { Spinner } from '@/shared/ui/spinner';
 export const FavoritesPage = () => {
   const dispatch = useAppDispatch();
 
-  const offers = useAppSelector((state) => state.offers);
-  const offersFetchStatus = useAppSelector((state) => state.offersFetchStatus);
+  const offers = useAppSelector(getOffers);
+  const offersFetchStatus = useAppSelector(getOffersFetchStatus);
 
   useEffect(() => {
     dispatch(fetchFavoriteOffers());
